+++ app/build.gradle
@@ -59,4 +59,6 @@ dependencies {
     implementation "com.squareup.retrofit2:converter-gson:2.6.0"
     implementation "com.squareup.okhttp3:logging-interceptor:3.12.6"
 
+    //Koin
+    implementation "org.koin:koin-android:2.0.1"
 }
