+++ app/src/main/java/com/sakal_n/myapplicationcinemasearch/viewmodel/HomeFragmentViewModel.kt
@@ -2,14 +2,16 @@ package com.sakal_n.myapplicationcinemasearch.viewmodel
 
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import com.sakal_n.myapplicationcinemasearch.App
 import com.sakal_n.myapplicationcinemasearch.domain.Film
 import com.sakal_n.myapplicationcinemasearch.domain.Interactor
+import org.koin.core.KoinComponent
+import org.koin.core.inject
 
-class HomeFragmentViewModel : ViewModel() {
-    val filmsListLiveData:  MutableLiveData<List<Film>> = MutableLiveData()
+class HomeFragmentViewModel : ViewModel(), KoinComponent {
+    val filmsListLiveData: MutableLiveData<List<Film>> = MutableLiveData()
+
     //Инициализируем интерактор
-    private var interactor: Interactor = App.instance.interactor
+    private val interactor: Interactor by inject()
 
     init {
         interactor.getFilmsFromApi(1, object : ApiCallback {
